fix(dashboard): move idea components to components dir
