+++ src/components/IdeaListItem/index.tsx
